Update loginUser controller.
Hi, I found a bug in the backend where you send a cookie even though the password is not correct. so when running loginStatus you will accidentally have a token. so the user can have access to the dashboard without actually logging in to the correct credentials, if they manually change the website URL.