Update EditProfile.js
I Think the save profile should not be included inside the if condition. This way, the user who for example just wants to change the name but doesn't want to change the photo, he wont be able to change the name either. I think the save profile section should come outside the if condition